* master: (96 commits)
  clarify CHANGELOG [ci skip].
  Fix Generation of proper migration when   ActiveRecord::Base.pluralize_table_names = false.
  update comments to reflect that options support is not available
  synchronize changelogs and 4.1 release notes. [ci skip]
  do not rely on method_missing hitting arel
  use ARel factory methods for building AST nodes
  Fix date_select option overwriting html classes
  - Rename `increment_or_decrement` to an apt `set_cache_value` since it actually doesn't increment/decrement in localstore.
  Check if any sqlite files are not included in the gitignore
  Remove sqlite3 lines from .gitignore if the application is not using sqlite3.
  Adding active_model in Rails::Info
  Clean up tables after each test.
  Swapped parameters of assert_equal in assert_select
  Update test helper to use latest Digestor API
  Digestor should just rely on the finder to know about the format and the variant -- trying to pass it back in makes a mess of things (oh, and doesnt work)
  Log the full path, including variant, that the digestor is trying to find
  Fix for digestor to consider variants for partials -- this still needs more testing!!
  fix log_tags request object grammar
  Extract with_example_table into helper method.
  test for structure:dump without schema information table. refs eafec46
  ...

Conflicts:
	activerecord/test/cases/relation/where_chain_test.rb

+++ activerecord/lib/active_record/railties/databases.rake
@@ -268,7 +268,8 @@ db_namespace = namespace :db do
       current_config = ActiveRecord::Tasks::DatabaseTasks.current_config
       ActiveRecord::Tasks::DatabaseTasks.structure_dump(current_config, filename)
 
-      if ActiveRecord::Base.connection.supports_migrations?
+      if ActiveRecord::Base.connection.supports_migrations? &&
+          ActiveRecord::SchemaMigration.table_exists?
         File.open(filename, "a") do |f|
           f.puts ActiveRecord::Base.connection.dump_schema_information
           f.print "\n"
